+++ lua_service/engine/memory_pool/MemoryPoolMgr.h
@@ -8,12 +8,12 @@ class MemoryPool;
 class MemoryPoolMgr
 {
 public:
-	MemoryPoolMgr(std::vector<uint32_t> block_sizes, uint32_t memory_page_size, uint32_t expect_working_block_set_num, uint32_t min_block_num_per_block_set);
+	MemoryPoolMgr(std::vector<size_t> block_sizes, size_t memory_page_size, size_t expect_working_block_set_num, size_t min_block_num_per_block_set);
 	~MemoryPoolMgr();
 
-	const static uint32_t BLOCK_SIZE_DESCRIPT_LEN = sizeof(uint32_t);
-	void * Malloc(uint32_t malloc_size);
-	void * Realloc(void *ptr, uint32_t new_size);
+	const static uint32_t BLOCK_SIZE_DESCRIPT_LEN = sizeof(size_t);
+	void * Malloc(size_t malloc_size);
+	void * Realloc(void *ptr, size_t new_size);
 	void Free(void *ptr);
 
 private:
@@ -25,10 +25,10 @@ private:
 	};
 
 	std::vector<MemoryPoolData *> m_memory_pools;
-	uint32_t m_max_block_size = 0;
+	size_t m_max_block_size = 0;
 	MemoryPoolData **m_memory_pool_fast_idx = nullptr;
 
-	const static int BLOCK_SIZE_MULTI_BASE = 8; // block must be times of 8
+	const static size_t BLOCK_SIZE_MULTI_BASE = 8; // block must be times of 8
 
-	static uint32_t CalPoolIndex(uint32_t malloc_size);
+	static size_t CalPoolIndex(size_t malloc_size);
 };
